created more code splitting for environments
